Attributes have now been fully extracted/moved from Enemies service to become its own Attributes service.
